+++ playlist_manager.js
@@ -1,5 +1,5 @@
 ﻿'use strict';
-//29/06/23
+//02/07/23
 
 /* 	Playlist Manager
 	Manager for Playlists Files and Auto-Playlists. Shows a virtual list of all playlists files within a configured folder (playlistPath).
@@ -422,7 +422,8 @@ if (!list.properties.bSetup[1]) {
 			const z = list.offset + Math.round(list.rows / 2 - 1);
 			list.cacheLastPosition(z);
 			list.update(false, true, z);
-			list.filter(); // Updates with current filter (instead of showing all files when something changes) and maintains focus on last selected item
+			// Updates with current filter (instead of showing all files when something changes) and maintains focus on last selected item
+			list.filter(list.searchInput && list.searchInput.text.length ? list.search(false) : void(0));
 			setTimeout(() => {if (pop.isEnabled()) {pop.disable(true);}}, 500);
 			return true;
 		}